src/requirements.txt
autopep8==1.5.5
cffi==1.14.5
pycodestyle==2.6.0
pycparser==2.20
PyNaCl==1.4.0
six==1.15.0
toml==0.10.2
